:recycle: Separate groom animation from state
